lecture-6-dependecy-injector-and-solid/d_injector_demo/requirements.txt
Django~=4.2.10
djangorestframework~=3.14.0